feat: floating point intervals

+++ integral_solver.c
@@ -1,11 +1,11 @@
 #include "integral_solver.h"
 
-double integral_solver(double (*function)(double), int a, int b, int n) {
+double integral_solver(double (*function)(double), double a, double b, int n) {
   if (n <= 0) {
     return 0;
   }
 
-  double step = (double)(b - a) / n;
+  double step = (b - a) / n;
   double res = (function(a) + function(b)) / 2;
 
   for (int i = 1; i < n; i++) {
